初步添加抽象语法树

+++ lib/github.com/GoExpress/eval_test.go
@@ -11,17 +11,17 @@ func TestEval(t *testing.T) {
 }
 
 func TestEval_number(t *testing.T) {
-	var aa=2
+	var aa = 2
 	fmt.Println(Eval(Add, &aa, 1))
 }
 
 func BenchmarkEval(b *testing.B) {
 	b.StopTimer()
-   var aa=1
+	var aa = 1
 	b.StartTimer()
 	for i := 0; i < b.N; i++ {
 		// var b,_=
-		 	Eval(Ride, &aa, 1)
+		Eval(Ride, &aa, 1)
 		//fmt.Println(b)
 	}
 }
@@ -29,12 +29,13 @@ func BenchmarkEval(b *testing.B) {
 type S struct {
 	Name *string
 }
+
 func TestEvalTake(t *testing.T) {
-	var aa=2
+	var aa = 2
 	fmt.Println(EvalTake("", &aa))
 
-	var s=S{
-		Name:nil,
+	var s = S{
+		Name: nil,
 	}
 	fmt.Println(EvalTake("s.Name", &s))
 }
@@ -42,8 +43,8 @@ func TestEvalTake(t *testing.T) {
 func BenchmarkEvalTake(b *testing.B) {
 	b.StopTimer()
 	//var aa=1
-	var s=S{
-		Name:nil,
+	var s = S{
+		Name: nil,
 	}
 	b.StartTimer()
 	for i := 0; i < b.N; i++ {
@@ -56,12 +57,12 @@ func BenchmarkEvalTake(b *testing.B) {
 
 func BenchmarkGetDeepValue(b *testing.B) {
 	b.StopTimer()
-	var a1=1
-    var v=reflect.ValueOf(&a1)
-    var bb *int
-    bb = &a1
+	var a1 = int64(1)
+	var v = reflect.ValueOf(&a1)
+	var bb *int64
+	bb = &a1
 	b.StartTimer()
 	for i := 0; i < b.N; i++ {
-		 GetDeepValue(v,bb)
+		GetDeepValue(v, bb)
 	}
 }
